Evaluate Reader or State effect in terms of base effect
If the effect stack is built on top a monad that has
MonadReader/MonadState capabilities, then it is easier to evaluate the
Reader/State effect directly in terms of base effect. For this purpose
we now provide following functions:

  runReaderAsBase
      :: forall m effs e a
      .  BaseMember m effs
      => m e
      -> Eff (Reader e ': effs) a
      -> Eff effs a

  runReaderM
      :: forall e m effs a
      .  (MonadReader e m, BaseMember m effs)
      => Eff (Reader e ': effs) a
      -> Eff effs a

  runStateAsBase
      :: forall s m effs a
      .  BaseMember m effs
      => m s
      -> (s -> m ())
      -> Eff (State s ': effs) a
      -> Eff effs a

  runStateM
      :: forall s m effs a
      .  (MonadState s m, BaseMember m effs)
      => Eff (State s ': effs) a
      -> Eff effs a

Unfortunately to implement runState{AsBase,M} we need unsafeCoerce,
because State effect data constructors aren't exported by freer package.